# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do sistema Estoque Fácil
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  code: string;
  name: string;
  /** @nullable */
  description: string | null;
  salePrice: number;
  /** @nullable */
  costPrice: number | null;
  currentStock: number;
  minimumStock: number;
  active: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  /**
     * @minLength 1
     * @maxLength 50
     */
  code: string;
  /**
     * @minLength 1
     * @maxLength 150
     */
  name: string;
  /**
     * @maxLength 500
     * @nullable
     */
  description?: string | null;
  /** @minimum 0 */
  salePrice: number;
  /**
     * @minimum 0
     * @nullable
     */
  costPrice?: number | null;
  /** @minimum 0 */
  initialStock?: number;
  /** @minimum 0 */
  minimumStock: number;
}

export interface ProductUpdate {
  /**
     * @minLength 1
     * @maxLength 50
     */
  code?: string;
  /**
     * @minLength 1
     * @maxLength 150
     */
  name?: string;
  /**
     * @maxLength 500
     * @nullable
     */
  description?: string | null;
  /** @minimum 0 */
  salePrice?: number;
  /**
     * @minimum 0
     * @nullable
     */
  costPrice?: number | null;
  /** @minimum 0 */
  minimumStock?: number;
}

export interface ProductStatusInput {
  active: boolean;
}

export type MovementType = typeof MovementType[keyof typeof MovementType];


export const MovementType = {
  ENTRY: 'ENTRY',
  EXIT: 'EXIT',
} as const;

export interface MovementInput {
  /** @minimum 1 */
  productId: number;
  /** @minimum 1 */
  quantity: number;
  /**
     * @minimum 0
     * @nullable
     */
  unitPrice?: number | null;
  movementDate: string;
  /**
     * @maxLength 100
     * @nullable
     */
  reason?: string | null;
  /**
     * @maxLength 500
     * @nullable
     */
  notes?: string | null;
}

export interface Movement {
  id: number;
  productId: number;
  productName: string;
  type: MovementType;
  quantity: number;
  /** @nullable */
  unitPrice: number | null;
  stockBefore: number;
  stockAfter: number;
  movementDate: string;
  /** @nullable */
  reason: string | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface DashboardSummary {
  activeProducts: number;
  totalItems: number;
  lowStockProducts: number;
  movementsThisMonth: number;
  stockValue: number;
}

export type StockReportItemStatus = typeof StockReportItemStatus[keyof typeof StockReportItemStatus];


export const StockReportItemStatus = {
  OK: 'OK',
  LOW: 'LOW',
  OUT: 'OUT',
} as const;

export interface StockReportItem {
  productId: number;
  code: string;
  name: string;
  currentStock: number;
  minimumStock: number;
  salePrice: number;
  stockValue: number;
  status: StockReportItemStatus;
}

export interface MovementReport {
  entries: number;
  exits: number;
  totalEntryItems: number;
  totalExitItems: number;
}

export interface Error {
  error: string;
}

/**
 * Dados inválidos
 */
export type BadRequestResponse = Error;

/**
 * Recurso não encontrado
 */
export type NotFoundResponse = Error;

/**
 * Operação não permitida
 */
export type ConflictResponse = Error;

export type ProductIdParameter = number;

export type SearchParameter = string;

export type LowStockParameter = boolean;

export type ActiveParameter = boolean;

export type MovementTypeParameter = MovementType;

export type FromParameter = string;

export type ToParameter = string;

export type LimitParameter = number;

export type ListProductsParams = {
search?: SearchParameter;
lowStock?: LowStockParameter;
active?: ActiveParameter;
};

export type ListMovementsParams = {
/**
 * @minimum 1
 */
productId?: ProductIdParameter;
type?: MovementTypeParameter;
from?: FromParameter;
to?: ToParameter;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: LimitParameter;
};

export type GetRecentMovementsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: LimitParameter;
};

export type GetMovementReportParams = {
from?: FromParameter;
to?: ToParameter;
};