# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do sistema Estoque Fácil
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  activeProducts: number;
  totalItems: number;
  lowStockProducts: number;
  movementsThisMonth: number;
  stockValue: number;
}