# lib/api-zod/src/generated/types/movement.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do sistema Estoque Fácil
 * OpenAPI spec version: 0.1.0
 */
import type { MovementType } from './movementType';

export interface Movement {
  id: number;
  productId: number;
  productName: string;
  type: MovementType;
  quantity: number;
  /** @nullable */
  unitPrice: number | null;
  stockBefore: number;
  stockAfter: number;
  movementDate: Date;
  /** @nullable */
  reason: string | null;
  /** @nullable */
  notes: string | null;
  createdAt: Date;
}