# lib/api-zod/src/generated/types/movementInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do sistema Estoque Fácil
 * OpenAPI spec version: 0.1.0
 */

export interface MovementInput {
  /** @minimum 1 */
  productId: number;
  /** @minimum 1 */
  quantity: number;
  /**
     * @minimum 0
     * @nullable
     */
  unitPrice?: number | null;
  movementDate: Date;
  /**
     * @maxLength 100
     * @nullable
     */
  reason?: string | null;
  /**
     * @maxLength 500
     * @nullable
     */
  notes?: string | null;
}