# lib/api-zod/src/generated/types/movementReport.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do sistema Estoque Fácil
 * OpenAPI spec version: 0.1.0
 */

export interface MovementReport {
  entries: number;
  exits: number;
  totalEntryItems: number;
  totalExitItems: number;
}