# lib/api-zod/src/generated/types/movementType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do sistema Estoque Fácil
 * OpenAPI spec version: 0.1.0
 */

export type MovementType = typeof MovementType[keyof typeof MovementType];


export const MovementType = {
  ENTRY: 'ENTRY',
  EXIT: 'EXIT',
} as const;