# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do sistema Estoque Fácil
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  code: string;
  name: string;
  /** @nullable */
  description: string | null;
  salePrice: number;
  /** @nullable */
  costPrice: number | null;
  currentStock: number;
  minimumStock: number;
  active: boolean;
  createdAt: Date;
  updatedAt: Date;
}