# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do sistema Estoque Fácil
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /**
     * @minLength 1
     * @maxLength 50
     */
  code: string;
  /**
     * @minLength 1
     * @maxLength 150
     */
  name: string;
  /**
     * @maxLength 500
     * @nullable
     */
  description?: string | null;
  /** @minimum 0 */
  salePrice: number;
  /**
     * @minimum 0
     * @nullable
     */
  costPrice?: number | null;
  /** @minimum 0 */
  initialStock?: number;
  /** @minimum 0 */
  minimumStock: number;
}