# lib/api-zod/src/generated/types/stockReportItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API do sistema Estoque Fácil
 * OpenAPI spec version: 0.1.0
 */
import type { StockReportItemStatus } from './stockReportItemStatus';

export interface StockReportItem {
  productId: number;
  code: string;
  name: string;
  currentStock: number;
  minimumStock: number;
  salePrice: number;
  stockValue: number;
  status: StockReportItemStatus;
}